pyproject.toml
[tool.poetry]
name = "first-server"
version = "0.1.0"
description = ""
authors = ["Ömer Gökhan Mert <omer0204@stud.kea.dk>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
requests = "^2.31.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
